+++ source/Internal/Framework/Module/Setup/Handler/ModuleConfigurationHandlerInterface.php
@@ -1,4 +1,5 @@
-<?php declare(strict_types=1);
+<?php
+
 /**
  * Copyright © OXID eSales AG. All rights reserved.
  * See LICENSE file for license details.