+++ source/Internal/Framework/Module/Setup/Handler/SettingModuleSettingHandler.php
@@ -1,8 +1,11 @@
-<?php declare(strict_types=1);
+<?php
+
 /**
  * Copyright © OXID eSales AG. All rights reserved.
  * See LICENSE file for license details.
  */
+
+declare(strict_types=1);
 
 namespace OxidEsales\EshopCommunity\Internal\Framework\Module\Setup\Handler;
 
@@ -11,9 +14,7 @@ use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject
 
 class SettingModuleSettingHandler implements ModuleConfigurationHandlerInterface
 {
-    /**
-     * @var SettingDaoInterface
-     */
+    /** @var SettingDaoInterface */
     private $settingDao;
 
     /**