+++ tests/Unit/Internal/Framework/Module/Setup/Handler/ClassExtensionsModuleSettingHandlerTest.php
@@ -1,24 +1,27 @@
-<?php declare(strict_types=1);
+<?php
+
 /**
  * Copyright © OXID eSales AG. All rights reserved.
  * See LICENSE file for license details.
  */
+
+declare(strict_types=1);
 
 namespace OxidEsales\EshopCommunity\Tests\Unit\Internal\Framework\Module\Setup\Handler;
 
 use OxidEsales\EshopCommunity\Internal\Framework\Config\Dao\ShopConfigurationSettingDaoInterface;
 use OxidEsales\EshopCommunity\Internal\Framework\Config\DataObject\ShopConfigurationSetting;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject\ModuleConfiguration;
+use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject\ModuleConfiguration\ClassExtension;
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Setup\Handler\ShopConfigurationClassExtensionsHandler;
 use PHPUnit\Framework\TestCase;
-use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject\ModuleConfiguration\ClassExtension;
 
 /**
  * @internal
  */
-class ClassExtensionsModuleSettingHandlerTest extends TestCase
+final class ClassExtensionsModuleSettingHandlerTest extends TestCase
 {
-    public function testHandlingOnModuleActivation()
+    public function testHandlingOnModuleActivation(): void
     {
         $shopConfigurationSettingBeforeHandling = new ShopConfigurationSetting();
         $shopConfigurationSettingBeforeHandling
@@ -59,7 +62,7 @@ class ClassExtensionsModuleSettingHandlerTest extends TestCase
         $handler->handleOnModuleActivation($moduleConfiguration, 1);
     }
 
-    public function testHandlingOnModuleDeactivation()
+    public function testHandlingOnModuleDeactivation(): void
     {
         $shopConfigurationSettingBeforeHandling = new ShopConfigurationSetting();
         $shopConfigurationSettingBeforeHandling