+++ tests/Unit/Internal/Framework/Module/Setup/Handler/ModuleVersionHandlerTest.php
@@ -1,8 +1,11 @@
-<?php declare(strict_types=1);
+<?php
+
 /**
  * Copyright © OXID eSales AG. All rights reserved.
  * See LICENSE file for license details.
  */
+
+declare(strict_types=1);
 
 namespace OxidEsales\EshopCommunity\Tests\Unit\Internal\Framework\Module\Setup\Handler;
 
@@ -13,9 +16,9 @@ use OxidEsales\EshopCommunity\Internal\Framework\Module\Configuration\DataObject
 use OxidEsales\EshopCommunity\Internal\Framework\Module\Setup\Handler\ModuleVersionHandler;
 use PHPUnit\Framework\TestCase;
 
-class ModuleVersionHandlerTest extends TestCase
+final class ModuleVersionHandlerTest extends TestCase
 {
-    public function testHandleOnModuleActivation()
+    public function testHandleOnModuleActivation(): void
     {
         $shopConfigurationSettingDao = $this->getTestShopConfigurationSettingDao();
 
@@ -34,7 +37,7 @@ class ModuleVersionHandlerTest extends TestCase
         );
     }
 
-    public function testHandleOnModuleDeactivation()
+    public function testHandleOnModuleDeactivation(): void
     {
         $shopConfigurationSettingDao = $this->getTestShopConfigurationSettingDao();